+++ backend/prisma/migrations/20260917003045_add_map_fog_enabled/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "Map" ADD COLUMN     "fogEnabled" BOOLEAN NOT NULL DEFAULT true;
+